macos: Stop passing -sdk to xcodebuild
The following OpenAFS macOS platform-specific components are built
with 'xcodebuild --sdk ${OSXSDK} --project ...':

  src/platform/DARWIN/
    AFSPreference/OpenAFS.xcodeproj
    AklogAuthPlugin/AklogAuthPlugin.xcodeproj
    afscell/afscell.xcodeproj

The value of OSXSDK is hardcoded at configure time (in
src/cf/sysname.m4) for each major macOS version and platform.

However, in recent years, this hardcoded approach has become unsuitable
because SDK versions are more likely to change their minor number when
Xcode is updated for a minor version macOS update.  For example, Xcode
versions intended for macos 14 (Sonoma, AFS_DARWIN230_ENV) have
increased the SDK version several times:

   macosx14.0
   macosx14.2
   macosx14.4
   macosx14.5

A similar situation has been observed for most recent versions of macOS,
at least back to 10.14 Mojave.

If the hardcoded OSXSDK version is not correct, xcodebuild will fail
with (for example, on 14 Sonoma):
   xcodebuild: error: SDK "macosx14.0" cannot be located.
   make[4]: *** [OpenAFS.prefPane] Error 64

To fix this, just don't pass -sdk at all, and xcodebuild will use
whatever sdk is installed (or, if there are multiple sdks installed, the
sdk that was specified as the default with xcode-select). Remove the
OSXSDK variable, since that's all it's used for.

At least for now, keep the existing method of hard-coding the sdk for
macOS 10.13 (AFS_DARWIN170_ENV) and earlier, to preserve existing
behavior. Use the new variable XCODEBUILD_FLAGS to specify the sdk on
these versions.

[adeason@sinenomine.net: Remove OSXSDK, introduce XCODEBUILD_FLAGS.]

Change-Id: I2e1b3dbb75f2600da2acd367ead960ab9c203e0c
Reviewed-on: https://gerrit.openafs.org/16150
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Ben Huntsman <ben@huntsmans.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>