tests: Make src/tests buildable
Several files in src/tests fail to build due to compiler warnings when
building with --enable-checking.  There are also unresolved references
during the link steps.

The warnings are for

    implicit-fallthrough
    unused-result
    unused-but-set-variable
    invalid-source-encoding

In addition the "all" make target references a variable that is set
later in the file, so the resulting list of targets is incorrect and are
not built.

Move libopr.a, ${LIB_hcrypto} and ${LIB_roken} from INT_LIBS into
COMMON_LIBS to resolve link failures in:
    test-setgroups
    test-setpag
    dup2-and-unlog
    create-stat
    rm-rf
    write-closed2
    afscp

Update code that prints usage to use warnx instead of relying on
__progname.

Check for errors for various syscalls, to avoid "unused-result"
warnings.

Remove 'read_buf1' in read-write.c to avoid an "unused-but-set-variable"
warning.

Add a missing 'break' to a switch statement in utime-file.c to avoid an
"implicit-fallthrough" warning. Add an AFS_FALLTHROUGH to fsx.c to avoid
the same warning; it seems like this may be a mistake and maybe should
be a "break", but match the existing fsx.c behavior at least for now,
since the intended behavior isn't completely clear.

Relocate the Makefile "all" target so it follows the macro definition
that it references.

Rename the top level Makefile target from tests to srctests for the
recipe that invokes make to build src/tests. Add the necessary
dependencies to the srctests target. Add srctests to the
finale_notest/finale_nolibafs_notest targets

Make sure the 'OpenAFS' dir exists when generating OpenAFS/Dirpath.pm
and OpenAFS/Dirpath.sh. For objdir builds, it may not exist when we
try to generate these files.

Note The code within the src/tests directory may be out of date. This
commit only ensures that the src/tests can build cleanly.  Further work
will be needed to validate the tests and ensure that they are
still working as intended.

Change-Id: Ic6d685c0dde5df25da175fe4516ce28a26443533
Reviewed-on: https://gerrit.openafs.org/15342
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>