AIX: Detect IBM Open XL C 17.1+ (clang) on AIX 7.2+
AIX 7.2 TL5 SP3 (7200-05-03) and higher support IBM's new clang-based
Open XL C 17.1 and up.  These OS releases continue to support the
traditional XL C 16.1 and below.  The older compiler is still the
default compiler for the AIX kernel and the 16.1 rte is shipped with
the base OS.

Both compilers can be installed simultaneously.  When compiling only
one or the other will be in $PATH.  This change allows the configure
script to act appropriately based upon which compiler is being used.

Change-Id: I9f6130a574f9c48e3bbd961266f0beb5646421e3
Reviewed-on: https://gerrit.openafs.org/15436
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>