AIX: Declare syscall()
In various places, we make direct libafs syscalls. On platforms with a
dedicated libafs syscall number, this involves passing AFS_SYSCALL to
a generic OS-provided function that issues the given syscall (usually
called syscall()).

On AIX, there is no generic syscall() function for issuing system
calls. Instead, system calls are treated like function calls, and are
translated into system calls at link-time. So our calls to e.g.
lpioctl() and lsetpag() are issuing system calls on AIX, and we have
no userspace definition of these functions.

Along with lpioctl() and lsetpag(), libafs also defines a system call
confusingly named 'syscall'. Since this is our system call, of course
the system headers provide no declaration for it, and so this causes
errors when building using the clang-based xlc 17.1, such as:

        .../src/volser/volmain.c:195:2: error: call to undeclared function 'syscall'; ISO C99 and later do not support implicit function declarations [-Wimplicit-function-declaration]
                syscall(AFS_SYSCALL /* AFS_SYSCALL */ , 28 /* AFSCALL_CALL */ , a3,
                ^

To avoid these errors, declare a prototype for syscall(). But be clear
that this is a function (technically a system call) provided by
OpenAFS, not by the OS.

Ideally our prototype would match the definition of syscall() in
afs_syscall.c, but our various callers pass a varying number of
arguments (as is normal for syscall() on other platforms). So at least
for now, declare it as a function with unspecified args.

Reviewed-on: https://gerrit.openafs.org/15446
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
(cherry picked from commit 9c4ce09091a953a7062e8e451c0a4b5d4393db51)

Change-Id: I5c4d51119feade1b3caab238daf87e1e5d4a2050
Reviewed-on: https://gerrit.openafs.org/15811
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>