AIX: Search harder for compiler path
The current logic for setting CC on AIX assumes the desired compiler
is in the user's PATH. But the desired compiler may not be in the
user's path, and we do not require it to be on other platforms.

So, if the user hasn't specified a CC explicitly, and we don't see a
compiler in the PATH, look in a few common known locations for one.
This is similar to our approach on Solaris, but the structure of the
known paths lets us easily use a shell glob, instead of needing to
specify every single compiler version.

Change-Id: I2cd16ab8bf9737edd52630d457de73546d89a19a
Reviewed-on: https://gerrit.openafs.org/15457
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Ben Huntsman <ben@huntsmans.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>