clang: Don't redefine printf in salvage.c
The Gentoo distribution changed the default fortify_source setting for
clang (to match the default being used for gcc).  This change causes the
following build error:

    src/dir/salvage.c:26:9: error: 'printf' macro redefined
        [-Werror,-Wmacro-redefined]
            ^
    /usr/include/bits/stdio2.h:89:11: note: previous definition is here
          ^
    1 error generated.

If the fortify_source level is greater than 1, glibc can define printf
as a macro.  The clang compiler has a default check for macro
redefinitions (-Wmacro-redefined), while gcc does not provide this
option.

Remove the:
   #define printf Log
in src/dir/salvage.c and update callers to use Log() instead of
printf().

Clean up lines that split the call to Log and its first parameter.  Fix
the indentation for Log's parameters.

There are no functional changes with this commit.

Change-Id: I732c8345dfd4cf140883bff4b53df9ab9bea5a85
Reviewed-on: https://gerrit.openafs.org/15462
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>