cf: Use static allocated structs for cf tests
The Linux kernel module build will issue a warning when a stack frame
exceeds a specific size via the -Wframe-larger-than= compiler flag (with
a default size of 2048 bytes on most architectures).

At least one distribution, Oracle's Linux with the Unbreakable
Enterprise Kernel (UEK), hardens this check by changing the warning to
an error (-Werror=frame-larger-than=).

Several of the openafs autoconf tests use objects that are allocated on
the stack when testing for structures, or members of a structure.
When the warning is changed to an error, configure fails in several
locations when testing against Linux's task_struct structure, which
exceeds 2048 bytes in size.

   openafs/conftest.dir/conftest.c:72:1: error: the frame size of 9984
   bytes is larger than 2048 bytes [-Werror=frame-larger-than=]

Update the autoconf Linux tests that allocate structures to use a
statically allocated structure instead of one allocated on the stack.

Change-Id: Ica44373b02fb4efa81d8fcbc0337aea9719b4fb6
Reviewed-on: https://gerrit.openafs.org/15477
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>