vos: Fix potential string buffer overflows
Several locations within vos use strcpy without verifying the length
of the source string which could lead to potential string buffer
overflows.

There are 2 cases where the string buffer can be replace with a pointer
to the string that was being copied.

Replace the string buffers with pointers where possible.
Replace the remaining uses of strcpy with strlcpy and and where
appropriate add checks to detect and report on possible string
truncations.

Change-Id: I4189f0ad1858065acf5bac7dfa83ce662e11fa4b
Reviewed-on: https://gerrit.openafs.org/15478
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>