Linux 6.5: Use register_sysctl()
The linux 6.5 commit:
    "sysctl: Remove register_sysctl_table" (b8cbc0855a)
removed the Linux function register_sysctl_table().  The replacement
function is register_sysctl(), which offers a simpler interface.

Add an autoconf test for the Linux function register_sysctl and add a
call to register_sysctl when available.

Notes:
The Linux function register_sysctl was added in Linux 3.3 with the
commit:
    'sysctl: Add register_sysctl for normal sysctl users' (fea478d410)
with a note that it is a simpler interface.

The function register_sysctl_table was marked as deprecated with the
Linux 6.3 commit:
    'proc_sysctl: enhance documentation' (1dc8689e4c)

Change-Id: I657bee103cd1f4ab2e60b3f3471c3fc12d8b6be4
Reviewed-on: https://gerrit.openafs.org/15500
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>