auth: Allocate ThisCell and CellAlias pathnames
The CellServDB pathname is dynamically allocated, however the ThisCell
and CellAlias pathnames are constructed with stack allocated buffers,
which limit the supported pathname lengths and makes this code path
susceptible to stack smashing.

Instead, use asprintf() to dynamically allocate the ThisCell and
CellAlias pathnames in order to support long configuration directory
paths.

Change-Id: I117a4109856ce58d875d7104a6b1805789fa55de
Reviewed-on: https://gerrit.openafs.org/15530
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>