linux: Replace fop iterate with fop iterate_shared
The Linux 6.5 commit:
  'vfs: get rid of old '->iterate' directory operation' (3e32715496)
removed the filesystem_operations iterate method.  The replacement
method, iterate_shared, was introduced with the Linux 4.6 commit:
  'introduce a parallel variant of ->iterate()' (6192269444)

The above commits indicate that the iterate_shared is an "almost"
drop-in replacement for iterate.  The vfs documentation for
iterate_shared has caveats on the implementation (serializing in-core
per-inode or per-dentry modifications and using d_alloc_parallel if
doing dcache pre-seeding).  A wrapper is provided to assist filesystems
with the migration from iterate to iterate_shared.  Until it can be
verified that afs_linux_readdir meets the above requirements, we will
use the wrapper (ref 3e32715496 commit)

Add configure tests for the iterate_shared file_operations member and
for the wrap_directory_iterator function.

Update osi_vnodeops.c to use iterate_shared and the wrapper if they are
both available.

Reviewed-on: https://gerrit.openafs.org/15528
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 7437f4d37719ea53711e06ac9675dad1abd6769e)

Change-Id: Id00cfab2c0b51c2167fe19cd9cf7f136450ff174
Reviewed-on: https://gerrit.openafs.org/15558
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>