linux: Remove use of wrap_directory_iterator
The commit 'linux: Replace fop iterate with fop iterate_shared'
(gerrit 15528) uses a Linux provided wrapper for the file_operations
iterate_shared member.  The wrapper is provided by Linux to give
filesystem time to migrate to using the iterate_shared directly.

Review of the locking done by the afs_linux_readdir shows that it is
unnecessary to use the migration wrapper.

Note: Removing the use of the wrapper as a separate commit from the
15528 commit in case there is a need to use the wrapper.  In addition,
we would like to use the wrapper with the current openafs stable release
and use the development branch for additional testing.

Change-Id: I662f95f69c2d06948f764b9520567efbca8f692c
Reviewed-on: https://gerrit.openafs.org/15550
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>