fsint: Move RXAFS_StoreACL function declaration
Commit 8f1eba056fd (CVE-2018-7168 RXAFS_StoreACL deprecate 134 introduce
164) added RXAFS opcode 164 as RXAFS_StoreACL. This change sensibly
added the new opcode 164 between opcodes 163 (RXAFS_DFSSymlink) and 220
(RXAFS_FsCmd) in the afsint.xg file.

Unfortunately, the rx rpc call statistics reported over the wire by
RXSTATS_RetrieveProcessRPCStats() are given by function index (not
opcode).  The function index is determined by the function position in
the xg file when rxgen generates the RPC stubs.  This means following
RXAFS RPCs stats are off-by-one in the returned stats block:

RXAFS_FsCmd
RXAFS_InlineBulkStatus
RXAFS_GiveUpAllCallBacks
RXAFS_GetCapabilities
RXAFS_CallBackRxConnAddr
RXAFS_GetStatistics64

Since the RPC call statistics are reported by function index and not
opcode, rxstat clients (e.g., rxstat_get_process) report incorrect stats
when there is a mismatch between the rxstat client and the fileserver.

Move the declaration of RXAFS_StoreACL (164) to be after the last
function in afsint.xg file, so the function indices generated by rxgen
are not changed for existing functions and the next available function
index is assigned to the new RXAFS_StoreACL (164) RPC.

Note: The RXSTATS_RetrieveProcessRPCStats() RPC has a parameter to
indicate the stat client version, which could be used in the future to
support an improved interface, since relying on the function indices is
not very robust.

Change-Id: Ic0adc8216fe80897600d13646588c3f27485e129
Reviewed-on: https://gerrit.openafs.org/15572
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>